+++ app/app.module.ts
@@ -6,24 +6,24 @@ import {MaterialModule} from '@angular/material';
 import {FlexLayoutModule} from '@angular/flex-layout';
 import 'hammerjs';
 import {CommonComponentsModule} from './components/components.module';
-import {RowSingleSelectComponent} from './components/table/row-single-select/row-single-select';
-import {RowActionMenuComponent} from './components/table/row-action-menu/row-action-menu';
+import {RowSingleSelectComponent} from './components/table/row/row-single-select.component';
+import {RowActionMenuComponent} from './components/table/row/row-action-menu.component';
 import {Routes, RouterModule} from '@angular/router';
 import {APP_BASE_HREF} from '@angular/common';
 
-import {NumericEditorComponent} from './components/table/editors/numeric-editor';
-import {SelectEditorComponent} from './components/table/editors/select-editor.component';
-import {TextEditorComponent} from './components/table/editors/text-editor';
+import {NumericEditorComponent} from './components/table/column/editors/numeric-editor.component';
+import {SelectEditorComponent} from './components/table/column/editors/select-editor.component';
+import {TextEditorComponent} from './components/table/column/editors/text-editor.component';
 
 import {ComponentPlaygroundComponent} from './examples/component-playground/component-playground.component'
 import {FooterRowGridComponent} from './examples/footer-row-grid/footer-row-grid.component';
 import {SingleSelectGridComponent} from './examples/single-select-grid/single-select-grid.component';
 import {MultiGridScreenComponent} from './examples/multi-grid-screen/multi-grid-screen.component';
 import {MultiSelectGridComponent} from './examples/multi-select-grid/multi-select-grid.component';
-import {DateCellRendererComponent} from "./components/table/cell/render/date-cell-renderer.component";
-import {DefaultCellRendererComponent} from "./components/table/cell/render/default-cell-renderer.component";
-import {ETCodeCellRendererComponent} from "./components/table/cell/render/et-code-cell-renderer.component";
-import {SelectEditorDirective} from "./components/table/editors/select-editor.directive";
+import {DateCellRendererComponent} from "./components/table/column/render/date-cell-renderer.component";
+import {DefaultCellRendererComponent} from "./components/table/column/render/default-cell-renderer.component";
+import {ETCodeCellRendererComponent} from "./components/table/column/render/et-code-cell-renderer.component";
+import {SelectEditorDirective} from "./components/table/column/editors/select-editor.directive";
 
 
 const appRoutes: Routes = [
