+++ app/components/table/row/row-footer-aggregation.factory.ts
