+++ app/components/table/row/row-single-select.component.tpl.html
