+++ app/components/table/row/row-single-select.component.ts
@@ -4,7 +4,7 @@ import {AgRendererComponent} from 'ag-grid-ng2';
 @Component({
   moduleId: module.id,
   selector: 'row-single-select',
-  templateUrl: 'row-single-select.tpl.html',
+  templateUrl: 'row-single-select.component.tpl.html',
   // tell angular we don't want view encapsulation, we don't want a shadow root
   encapsulation: ViewEncapsulation.None
 })
